test: Remove Services SCTP test case
The test case is supposed to test whether the BPF NAT engine doesn't
drop packets of a type which it doesn't support (e.g. SCTP). However,
recently the test case became flaky. Considering this and a little value
the test provides, let's remove the test case.

Instead, we should create a BPF unit test for it once it's available
(#16862).

Signed-off-by: Martynas Pumputis <m@lambda.lt>